Refactor
* move deletePackageCache to registry/index.ts
* clean up imports
